requirements/requirements-testing.txt
# Pytest for running the tests
pytest==2.9.1
pytest-django==2.9.1
